tests: Mark new bridge test as broken
In commit acc19386f66d (rpc: Add test for 'int' parameters in Bridge) we
added a new test, which fails consistently on the CI and locally.

The test adds a fake api with 3 arguments, while the schema specify only
2 arguments, so the new api is called with a missing argument, raising
InvalidCall.

Marked as broken for now as it breaks the build.

Change-Id: I2eb82e4dcf6afe2d4718fa3b185acf0b6fbbe869
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57364
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>