hostdev: fix scsi udev rule target
Previously, the rule worked as expected but did not properly signal
ownership of the device. That is now fixed with kernel and subsystem
targets.

The cause does not affect functionality of hostdev-scsi, but causes the
rule to be triggered with all udev updates - leading to useless chown
execution.

Change-Id: I98949e645b748493e3cf4a7377d17ded67e8fd66
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57373
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>