monitoring: Use new monitoring infrastructure
DomainMonitor is keeping now a checker object, providing path checking
services. The checker starts with the domain monitor is created, and
stopped when shutting down vdsm.

Monitor threads use the checker to start checking the monitoring path
during setupLoop, and stop checking when the thread is stopping.

Since path checker events are received on a different thread, creating
new status and updating it is done under a lock to races when evaluating
status changes.

For the same reason, we must check if the host id should be acquired
both when finishing domain status check and path status check. Both
checks run in different rate.

The logic for acquiring a host id was enhanced to support partial
status, which was not possible before.

Change-Id: I18c87f3be6c570ee914388fc2ef7d6b847516b7c
Bug-Url: https://bugzilla.redhat.com/1081962
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57432
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>