ifcfg: ad_actor_system can not be reset to its default value
_restore_default_bond_options tried to reset this option to
00:00:00:00:00:00, which (on Fedora 23) causes "IOError: [Errno 22]
Invalid argument", during testSetupNetworksActiveSlave.

Linux documentation (bonding.txt) says that "The value can not
be either NULL or Multicast".

Let's add ad_actor_system to EXCLUDED_BONDING_ENTRIES for now
and assume that if a user set ad_actor_system explicitly, she
knows the way to reset it, and that is by setting it to the bond's
own MAC address:

  In an AD system, this specifies the mac-address for the actor in
  protocol packet exchanges (LACPDUs). The value cannot be NULL or
  multicast. It is preferred to have the local-admin bit set for this
  mac but driver does not enforce it. If the value is not given then
  system defaults to using the masters' mac address as actors' system
  address.

Change-Id: I3f307b64546c3efa73ca202e97ced49eeb9670e9
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57490
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>