netinfo: improve the list of vlan devices on top of an interface.
The current code assumed that all links on top of an interface are vlan
links. Since the links could also be of an other type, e.g. veth, the
current implementation does not behave correctly for links of an other
type but vlan.

The improvement is that now we will check if link is of type vlan.

Change-Id: I11e217764125c1931a201f05863c5359fa66dbaf
Signed-off-by: Dominik Holler <dominik.holler@gmx.net>
Reviewed-on: https://gerrit.ovirt.org/57500
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>