virt: Collect vm start flags from device hooks when running vm
We must be able to pass more information from the
device hooks than just the device xml. This can be done
by placing additional flags which can be use to alter vm
behaviour in a config file, which will then be passed to
the vm config.

This is needed for example to solve a race condition
when nic managed by external providers are added, and
the vm boots using one of these nics. The nics are
only connected at startup, which takes a few seconds.
By then the vm already tries to boot, and fails because
of no network connection (as the nic is still not ready).
Adding the libvirt.VIR_DOMAIN_START_PAUSED flag will start the vm paused,
giving time to connect the nic. The vm can then be
resumed in the after_vm_start hook, once the nic is
in the up state.

Change-Id: I554e304027ea25340634e8def5ddeef99d98a10a
Bug-Url: https://bugzilla.redhat.com/1314371
Signed-off-by: mirecki <mmirecki@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57538
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58898
Reviewed-by: Dan Kenigsberg <danken@redhat.com>