Import storage constants as sc
We have two constants modules: vdsm.constants and
vdsm.storage.constants.  The storage constants must be imported with a
different name but storage_constants is too long given the amount of use
this name will see throughout the code.  Adopt the convention used for
storage_exceptions (se) and use sc as the name.  Do so uniformly to
increase readability.

Change-Id: If95bcc6779e4fee0bb084d7e7432e321e31bd95b
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57580
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Ala Hino <ahino@redhat.com>
Continuous-Integration: Jenkins CI