tests: Add monitor thread tests
Add tests before we refactor MonitorThread for new monitoring
infrastructure.

To allow testing refresh timeout, we parse is now as float instead of
int. This is backward compatible with old code but allow very short
timeouts during the tests.

These tests cover close to 100% of the MonitorThread class, which we
plan to modify heavily. We have no test for the DomainMonitor class, but
we do not plan any significant change in this class.

To make the test robust even on slow CI machines, a callback was added
to the monitor thread. If the callback is set, the monitor calls it once
after finishing a monitoring cycle. The tests use this callback to
synchronize with the monitor thread.

Change-Id: Ib7c6766435b593828f2851fa68c39aa584b4e9ad
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57585
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>