monitor: Pass change event to monitor threads
We used to pass the DomainMonitor object to MonitorThread, so the thread
can emit state change events using the domain monitor's
onStageChangeEvent instance variable. This couple together the objects
and make them harder to test for no benefit.

Now we pass the onStateChangeEvent object to the monitor threads so they
can emit events.

Change-Id: I48ddc78d0e0573ecaa5c29c621c0262f5e77d8c9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57706
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI