virt: Limit the number of workers in executor
In some situations, periodic operations may hang inside executor worker
threads.  We have observed that in real situations when libvirt calls
started being blocked due to some bug.  Such situations shouldn't happen
when everything works correctly, but actually not everything works
correctly and there are unexpected bugs.  We should try to minimize the
possible damages.

The result of the hanging periodic calls is that the number of worker
threads in the executor grows above all limits.  This is not good.  So
we limit the maximum number of the worker threads in the executor with
this patch.

We set the maximum number of the threads to a value sufficiently high
for harmless temporary problems.  If the number of the workers is higher
than that then there is probably a real problem and we should start
emergency actions, i.e. not spawning further threads in this case.

Note that a single isolated problem of this kind may block all the
periodic operations.  We introduce a bit more fine grained approach to
handle that in a followup patch.

Change-Id: Iba56d91474c6b14a1cfe2db827b6fd61843a1db2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Backport-To: 4.0
Backport-To: 3.6
Reviewed-on: https://gerrit.ovirt.org/57754
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI