migration: move helper into Progress class
move the update_progress() helper inside
the new migration.Progress class with no intended
changes in behaviour.

Additionally, now that we have the percentage
property in the Progress class, we can enhance its
__str__ magic method to be fully expressive, instead
of logging half of information in the
client code and half in the class itself.

Change-Id: If8e5a5cdf2249d73ef45c4f50e1fa6c4338ad8a0
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56313
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>