spec: Require kernel to solve bnx2fc corruption
kernel-3.10.0-327.18.1.el7 solves a corruption in the bnx2fc driver
used for FCoE. This patch requires the aforementioned kernel release
so that FCoE can be safely used on VDSM hosts (via the hook introduced
in commit 16cb9db).

Note that as this kernel is available for both RHEL and CentOS[1] 7.2,
there's no reason to maintain the separation between the two in the
spec file, and it may be cleaned up to have a single kernel
dependency for all the ELs.

[1] http://mirror.centos.org/centos/7/updates/x86_64/Packages/

Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1338543
Change-Id: I718a97bf390254b59e1530b5a5c20465ecec6ed2
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57849
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>