hostdev: improve robustness of libvirt SCSI access
SCSI host device is specific as it's identified by address and adapter.
Previously, we only cared about address which is good enough in most
of the cases except hotplug and hotunplug.

Since we can now work inside the device class without having to deal
with other devices, we can add adapter matching and correctly handle
the address.

The change implies removal of mutation in scsi_address_to_adapter. That
is intentional and should be done as a single patch, as trying to
separate is unsafe.

Change-Id: I0e9983fd309869ebd8a4aacd8bfe42158cd26e64
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58011
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>