tests: Fix flaky test
The test is trying to test request timeout, when the server response is
too slow. Because it was using very short timeout (0.01 seconds), it was
triggering deadlock the server.

To reproduce locally, I changed the timeout to 0.001 seconds. When the
process get stuck, we can see two threads:

Thread 2 (Thread 0x7fd77db42700 (LWP 30374)):
/usr/lib64/python2.7/site-packages/M2Crypto/SSL/Connection.py, line 217,
in _read_bio
    return m2.ssl_read(self.ssl, size, self._timeout)

Thread 1 (Thread 0x7fd785d1e700 (LWP 30369)):
/usr/lib64/python2.7/SocketServer.py, line 251, in shutdown
    self.__is_shut_down.wait()

Thread 2 is blocked on ssl_read that will never receive any data since
the client timed out, and the other thread is waiting on the server's
__is_shutdown event, which will never return.

This may be an issue in M2Crypto or in our testing server, but these
tests are for the client, not the server.  Fix the tests by increasing
the timeout on both the server and the client.

Change-Id: I44f2deb1040ec387da699a56ad7ed3e474888e87
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57993
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>