virt: Use libvirt to get drive size
Due to a bug in libvirt versions 0.10.2 and below, we use qemu-img to
discover the virtual size of a running VM's disk when live resizing it.
Since all supported platforms depend on a libvirt with the fix, we
should use the virDomainBlockInfo API to get the drive size.

Change-Id: I19b9f452bae0a1c03f2efe543a4bfa0d671e932d
Bug-Url: https://bugzilla.redhat.com/1128855
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55399
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI