systemd: Specify number of open file limit for vdsmd process
Specify for vdsmd process the limit for open files.
We have a specification in limits.conf, but it is setting limit
for the vdsm user not for the process.

The ulimit for the process can be checked in /proc/PID/limits
(where PID is the vdsm pid)

In commit I26345e 'BZ#769502 Move limit configuration to its proper
place', the configuration moved from setting the limit from the python
code using ulimit to limits.conf.

Change-Id: Ic8b4ddb54e812658e120f7ad10fba81dd8feef87
Bug-Url: https://bugzilla.redhat.com/1339245
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58123
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>