Revert "test: Introduce mock module for testing"
This reverts commit 252fcad9d57c80883121a8385ca678341d6c0c96.

This commit added python-mock as build dependency, but it seems that the
package is not available on el7.

The package is available in
http://mirror.centos.org/centos/7/virt/$basearch/ovirt-3.6/

But not in:
http://cbs.centos.org/repos/virt7-ovirt-40-candidate/x86_64/os/

This may be a repository issue, but we cannot required this for centos
until this issue is fixed.

It seem to be missing on RHEL, or maybe I'm missing the magic
repository holding it.

We can install this package using pip for automation, but we cannot
require it for building vdsm. This also means that we cannot run the
tests in make rpm on el7.

Change-Id: I32c6c375d0cace936e6246b7e367d843cd77a25b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58182
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>