virt: Return first child from vmxml.xml_to_dom
Generally we don't want to work with Document instances, we want to work with
Element instances.  There is no need to return Document instead of the top
Element from vmxml.xml_to_dom.  And etree behaves this way anyway, so we should
be ready for that change.  We just must be careful that getElementsByTagName
doesn't return the element it's applied on, but vmxml.dom_find_all handles
that (and again, the similar applies to etree).

parse_xml is used only in a few places and using Element instance instead of
Document is fine everywhere.

We need this change in the following patch converting vmxml.Domain to
DOM helpers.

Change-Id: I7e4c92c32e8fa88836d9434a3917b758feead277
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58194
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI