migration: downtime: start thread after 1st iter
This patch makes the migration.DowntimeThread start only after
the first QEMU iteration.

A "QEMU iteration" is the process on which QEMU, during a pre-copy
live migration, attempts to transfer all the memory pages from the source
side to the destination side.

It is better to start the downtime adjustment only after the first
iteration, because the very fact QEMU needs more iterations means the
migration is not converging (easily).

This patch ensures we still set up a default downtime using our
own configured values.

Change-Id: I661c44d9dd7bdb9f1ec3a355038cfa8966946011
Backport-To: 4.0
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58417
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>