rpc: Log important info from VM stats
Currently, we don't log results of getAllVmStats API calls.  This is not
to fill logs with a lot of uninteresting information.  But information
contained in the VM stats is useful and we'd like to have it somewhere.

A possible solution to these contradictory requirements is to log just
important information and to log it only occasionally.  In this patch,
we introduce logging VM ids and states in getAllVmStats calls from time
to time.

We implement this as a rather general functionality.  We are going to
use the same mechanism for other large API outputs in followup patches.

There may be performance concerns but please consider that we perform
data extraction only for calls that are global (not per VM) and that are
not very frequent (like every second or so).  So the processing overhead
shouldn't hurt much and shouldn't be much big in comparison to creating
the returned data inside Vdsm.

Change-Id: Ifcbac615323b62fb9a27e5c0f5a4e98990076146
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1339291
Backport-To: 4.0
Backport-To: 3.6
Reviewed-on: https://gerrit.ovirt.org/58465
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>