migration: downtime: handle steps=1
The migration.DowntimeThread() computes the sequence of values
of the downtime setting to be sent to libvirt.

The computation uses the (user-configurable) number of steps
as input value, and it has a special handling for steps == 1.

This patch makes the computation more uniform, in order
to prepare the land for a future patch, which wants to improve
how the DowntimeThread is used.

Change-Id: Ibb99e617d227e8043a1e99f0e8b1040e02a9b7e3
Backport-To: 4.0
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58416
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>