ovs: edit bond mode
With this patch we allow setup of OVS bond with one of 'fallback'
modes 1 and 4 and miimon.

In another patch we will allow users to edit all OVS bond parameters.
Fallback modes must be used for legacy->ovs migration. Bonds with other
modes or options cannot be migrated.

Change-Id: I9009933445c8bf162b236a2f9081e796d5f9272f
Bug-Url: https://bugzilla.redhat.com/1195208
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58437
Reviewed-by: Edward Haas <edwardh@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61112