api: Use warnings to report schema inconsistencies
Previously we spammed the log with warnings about conflicts between the
schema and the real usage of the apis. These warnings are useful for
developers but have negative value for users.

Now we use warnings.warn instead of logging.warning, so they do not
appear in the log by default. To enable warnings in the log, change the
configuration:

    [devel]
    python_warnings_enable = true

When enabled, the warnings are little more verbose now, using Python
standard warning style:

    jsonrpc.Executor/2::WARNING::2016-06-01
    22:23:55,551::vdsmapi::148::py.warnings::(_report_inconsistency)
    /usr/lib/python2.7/site-packages/api/vdsmapi.py:148: Inconsistency: No
    default value specified for systemProductName parameter in
    Host.getHardwareInfo
      warnings.warn(message, Inconsistency)

Change-Id: I5077d77b3b0e18edafa5f85073feb39cb87448cb
Bug-Url: https://bugzilla.redhat.com/1348256
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58494
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/59512
Reviewed-by: Francesco Romani <fromani@redhat.com>