tests: storageServerTests refactor
This patch changes the spec argument that is sent to MountConnection to
be of the right form. That it, instead of using "s" or "spec", we now
use "server:/path".

This patch does not fix the bug, but required for the next patch fixing
the bug.

Change-Id: Ic26860c9ebabc7eefe84d5fa5c705d5f9c2e73ff
Bug-Url: https://bugzilla.redhat.com/1305529
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58565
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/58680
Continuous-Integration: Francesco Romani <fromani@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>