logging: Fix duplicate ioprocess logs
Messages from the IOProcess logger are logged twice:

    ioprocess communication (4660)::INFO::2016-06-05
    15:52:07,240::__init__::434::IOProcess::(_processLogs) Starting
    ioprocess

    ioprocess communication (4660)::INFO::2016-06-05
    15:52:07,240::__init__::434::IOProcess::(_processLogs) Starting
    ioprocess

(Actual lines from vdsm log, wrapped and separated)

This happens because IOProcess logger uses propagate=True, but is
attached to the logfile handler. Python logs the message to the handler,
and then pass the message to the root logger, logging it again to the
same handler.

Change-Id: Ifd2ce2abf77f5688beec542c07aa124ab064976e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58644
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>