rng: properly teardown the device
The code is not correct (see fn signature) and does nothing except for
logging an error. The real issue is masked by generic
rmAppropriateMultipathRules function, cleaning the file for us if VM
has a disk.

This patch corrects the deletion itself, while the followup patch tries
to save hwrng files from being removed through storage code.

This patch is a trivial cherry-pick from master where it fixes a
different issue (see master' patch commit message reported verbatim
below). We choose to backport this patch because it fixes the issue
and it makes the code consistent with master branch, avoiding unneeded
differences.

== commit message of the master's branch patch ==

Patch 341a6cb5 moved the class functionality related to RNG sources to
separate module.
Unfortunately, we did not encounter git conflict when 9857a9c3 was
merged before the patch.
We fix this by using the new code as appropriate.

Change-Id: Ie4550da29457653c5f8ed45f9908ce2957f2f56b
Backport-To: 4.0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1343859
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58676
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58921
Tested-by: Francesco Romani <fromani@redhat.com>