jsonrpc: Fix log level overriding of some methods
In commit ba171f1d, checking of method names to filter their logging out
was changed.  We no longer check the mangled names, but we forgot to
update the list of method names to filter out accordingly.  This patch
fixes that.

Please note that not logging those method calls at all is not a good
idea, but we are going to fix that in followup patches.

Change-Id: Ib3e684b7fbe94aa4de11ceba86793da3aa4ad6b5
Bug-Url: https://bugzilla.redhat.com/1339291
Backport-To: 4.0
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58462
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58707
Continuous-Integration: Francesco Romani <fromani@redhat.com>