rpc: Lower logging priority just for getAllVmStats
Currently, we effectively disable logging of calls for several methods.
But we should consider whether it causes more harm to log the
corresponding information or not to log it.

We should avoid logging excessively large messages.  This applies to
getAllVmStats that may report quite a lot of information for many VMs.
But there is probably no good reason not to log everything about the
other methods currently silenced.  So with this patch we suppress
logging just of getAllVmStats.

We should be informed about that method as well, we handle that in
followup patches.

Change-Id: I5eefac9f88ede84df06c0cc39b188c8ae75e456b
Bug-Url: https://bugzilla.redhat.com/1351247
Backport-To: 4.0
Backport-To: 3.6
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58463
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/58708
Continuous-Integration: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59992
Reviewed-by: Nir Soffer <nsoffer@redhat.com>