virt: enableGuestEvents machine params needs to be set later
Previously the enableGuestEvents value got overwritten because
it was initialized too early. The setting of the value has been moved
to a later point to avoid such mistakes in the future.

The problem was caused by the fact that the 'enableGuestEvents' value
was kept in self.conf and not dropped, which was originally set to a
value on the first migration. That first migration overwrote the value
which was passed by the engine later.

This patch now drops the value from self.conf on the migration destination
and moved the value setting to a later point in time on the source host.
This should make the whole thing more resilient to such issues.

Change-Id: I43270faac2094068905d8d83c3fdee6851ff11d2
Backport-To: 4.0
Bug-Url: https://bugzilla.redhat.com/1344361
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58926
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit d35b8e795f7a7f7cac48f3da427996ec99223fc7)
Reviewed-on: https://gerrit.ovirt.org/59020
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>