tests: hooks: fix monkeypatching
The commit ee9eb95 rightfully added a test, with monkeypatching,
to avoid writing in system folders.

However, the path which is attempted to be mocked is created as
constant at import time, so the current monkeypatching happens too late,
making it useless.

This patch fixes that by using more aggressive monkeypatching: we
replace a private module constant.

Change-Id: Id921b25dd7cfc8965e1b56d2d2cbb2bc4bd36171
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58963
Continuous-Integration: Jenkins CI
Reviewed-by: Marcin Mirecki <mmirecki@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>