net tests: Support the ability to choose the tested switch
Use VDSM_TESTER_SWITCH_TYPE env variable to specify which switch type
should be tested in the functional tests.
If it is not set, all switch types will run.
If 'ovs' (or 'legacy') is set, then 'legacy' (or 'ovs') are skipped.

The option has been introduced to overcome the problem described by
https://bugzilla.redhat.com/show_bug.cgi?id=1346232

Change-Id: I06e230633dfefcdc4a4b545eb2139ffe190c6f35
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/58970
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>