rng: allow urandom as virtio rng entropy source
Libvirt previously used a whitelist to only allow random and hwrng
virtio-rng sources. As per libvirt commit
67f2b72723c242969c5282fcb9acf00cc01f2a54, this restriction is lifted
and any character device may be used. We don't exactly want to use
*any* character device, but we may allow usage of urandom. Additional
discussion in RHBZ#1074464.

The implementation hides the fact that urandom can be used.  This is
due to the fact that 4.0 (and lower) engine can't handle additional
RNG sources.

Change-Id: I2d46f1f3cff63e6443f5154577bd57b5ffc5f4ce
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1347669
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59031
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>