rpc: Use Suppressed class instead of logging workarounds
We have got logging workarounds for getAllVmStats RPC calls.
The workarounds are there to suppress logging large data regularly.
We'd like to get rid of the workarounds, it's better to handle the large
data in the API functions, where we can use specific handling of each
kind of large data.

This patch introduces a wrapper for large values, with the purpose of
suppressing writing them to the log.  The wrapper is used for
getAllVmStats so that we get rid of the logging workarounds.

There are other API calls that may produce large outputs, we will handle
them in followup patches.

Change-Id: Idaf00e557fccb8f08fa3aeb38d51cb4bbe0ffe53
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1351247
Backport-To: 4.0
Backport-To: 3.6
Reviewed-on: https://gerrit.ovirt.org/59078
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59994
Reviewed-by: Francesco Romani <fromani@redhat.com>