rpc: Log RPC call summary on info level
We currently log relatively detailed information about all RPC calls.

This information is logged on the debug level.  We are going to switch
to the info logging level sometimes in future and then information about
RPC calls won't be logged anymore.  But we still want to see some very
basic information about the RPC calls, such as that they were performed
and how much time they took.

This patch adds the corresponding logging.  We additionally add the
response to the context at a single place as a preparation for the
following patch that logs the response.

Change-Id: Idde2f1ba7394f16770543f5ca13411e8c2339cc6
Bug-Url: https://bugzilla.redhat.com/1351247
Backport-To: 4.0
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59080
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/59998
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>