infra: moved infra package from lib/vdsm/infra to lib/vdsm/common
infra package was under lib/vdsm/infra with a few sub packages.
every package was converted to module and replaced under lib/vdsm/common
vdsm-infra package was also removed from vdsm.spec.in and marked as
obsolete as it was redundant and there is no use for it right now.

Change-Id: I8ffd40750048371c58628dec5890b9f8f97d22b6
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59304
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>