API: streamline and make setLogLevel correct
According to the schema, setLogLevel should accept
a log level string ('DEBUG', 'INFO',...). But the code
actually blindly casted the given value to int(), and this suggests
the code actually expected an integer value.

To make things a bit user friendlier and comformant to schema,
change the argument to actually be a log level in string format.

Along the way, this patch streamlines the implementation of
setLogLevel and makes it simpler.

Change-Id: Iaddbb12d13bdbaa7a02255ab209da11e42d2ea97
Backport-To: 4.0
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1351077
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38425
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59873
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>