storage: add discard support for vm disks
When the engine sends the field 'discard' with the value true as
one of the parameters of the vm's disk's drive, vdsm adds
discard='unmap' to the disk's drive.
With that, vdsm gives the vm disk the ability to support discard.

Note that vdsm can handle scenarios of sending discard with the
values true or false, alongside not sending it at all, which makes
it possible to work with old engines that don't support this feature.

For more information about the "Pass Discard" feature, please refer to
the feature page:
http://www.ovirt.org/develop/release-management/features/storage/pass-discard-from-guest-to-underlying-storage/

Change-Id: I215c12260f819538e40056ec16d0b9378287ccee
Bug-Url: https://bugzilla.redhat.com/1241106
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65751
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI