jsonrpc: Pass JsonRpcInternalError instance, not class, to response
If there is a problem with processing a response in
_JsonRpcServeRequestContext.sendReply, the method generates a general
error response.  However, due to missing parentheses,
JsonRpcInternalError class, instead of JsonRpcInternalError instance, is
passed as `error' argument to JsonRpcResponse.  This leads to a crash
when JsonRpcResponse tries to access `error'.

This patch fixes the error by adding the missing parentheses.

Change-Id: I7c3d4065374d739bbecdbe4c17d7e513153c034a
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Backport-To: 4.0
Backport-To: 3.6
Reviewed-on: https://gerrit.ovirt.org/59355
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/59554
Reviewed-by: Francesco Romani <fromani@redhat.com>