logging: Introduce throttledlog
In some situations, we want to log interesting information.  A natural
place to retrieve and log the information is where it is created.
However the corresponding code may be called quite often and we don't
need logging the information that often and fill logs with it
excessively.  It's sufficient to log it just from time to time.

This patch provides means for occasional logging.  The provided
functions actually log only every given number of invocations or every
given time span.  We're going to use the functionality in followup
patches.

Change-Id: Ief96f74eb84880827ccf24c3e7ae854403090b8a
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1351247
Reviewed-on: https://gerrit.ovirt.org/59461
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59995
Reviewed-on: https://gerrit.ovirt.org/60100
Continuous-Integration: Francesco Romani <fromani@redhat.com>