migration: downtime: add tunable for back compat
In order to fix rhbz#1339521 and to improve the general
behaviour of the migrations, we backported
the changes
I661c44d9dd7bdb9f1ec3a355038cfa8966946011
and
I72ef8dc43e2143b7eb36f65ceb20f4b7ac919c5a
together with all their dependencies.

The patches change the way we update the VM migration
downtime, improving the algorithm to compute the
wait between steps, and starting to update the downtime
after one QEMU iteration[1] was done.

Those patches are suitable for a backport, but
they change the behaviour of the migrations in a way,
supposedly positive, that users could find surprising,
especially when we are so deep in the stabilization of
the branch (around 3.6.8).

We still believe that offering this options to the
users is the right call; however, with this patch we
add an option to disable the new migration downtime
adjustments, allowing the user to keep the
pre-3.6.8 behaviour.

+++

[1] a "QEMU iteration" is the process on which, during
a pre-copy migration, QEMU tries to transfer all the
dirty ram pages from the source VM to the destination
VM. Depending on the memory pressure on the source VM,
one migration may need more than one iteration
to succeed.

Change-Id: If477f268354a9609ecca216dad1eee4f40910c8f
Label: ovirt-3.6-only
Bug-Url: https://bugzilla.redhat.com/1339521
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59467
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI