vdsm: removed unnecessary calls to vdsm config.read
libvirt and vdscli modules had calls to config.read with vdsm default
path as a parameter. This is unnecessary because when the module is
imported the config is already read from the same path.

Also removed VDSM_CONF from libvirt FILES dictionary since it's not
needed anymore and fixed toolTests.py to match those changes.

Change-Id: I965d5adf0c2d5857109cfeca936c9a3139c111cb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1279555
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59513
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/59802