sp: allow executing upgradePool even if there is a pending update
Currently when a pool upgrade is initated (either by startSpm() or by
executing upgradePool() explicitly) the _domainsToUpgrade member of the
storage pool is initiated with the active domains, when each domain
upgrade is completed (or if the upgrade is unneeded) its id is removed
from that list.

When there is an unavailable domain, it'll remain on that list until its
update is complete - while it's unavailable other upgrade can't be
executed.
This patch changes that, we should allow upgrading the pool even if an
older upgrade wasn't complete yet - the domains that the old update
wasn't completed yet for will be upgraded to the new version.

Change-Id: I8e14a3aa33bfab4751ab5d1e3becbeda892da4c3
Bug-Url: https://bugzilla.redhat.com/1260428
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57188
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/60519
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Yaniv Bronhaim <ybronhei@redhat.com>